requirements.txt
beautifulsoup4==4.6.0
requests==2.18.4
python-telegram-bot==9.0.0
Django==2.0.2
django-telegrambot==1.0.1